requirements.txt
fastapi
uvicorn
requests
sqlalchemy
psycopg2-binary
APScheduler
python-dotenv
httpx
PyJWT
